- Improved component structure
