+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {withProviders} from "./system/providers/providers";
-import {AppContainerBox} from "./components/AppContainerBox";
+import {AppContainerBox} from "./components/AppContainerBox/AppContainerBox";
 
 const AppInit = withProviders(() => (<AppContainerBox />));
