+++ src/system/layouts/PageCenteredLayout.tsx
@@ -3,7 +3,7 @@ import styled from "@mui/material/styles/styled";
 import Box from "@mui/material/Box";
 import Paper from "@mui/material/Paper";
 import {PaddedContentFlexBoxColumn} from "../../components/common/components";
-import {AppHeader} from "../../components/AppHeader";
+import {AppHeader} from "../../components/AppHeader/AppHeader";
 
 const CenteredFullHeightFlex = styled(Box)(({theme}) => ({
   height: '100vh',
